# Conflicts:
#	.idea/misc.xml
#	src/main/java/Main.java
#	src/main/java/controleur/Controleur.java
#	src/main/java/modele/Player.java
#	src/main/java/vue/PlayerView.java
#	src/main/java/vue/TerrainView.java
#	target/classes/controleur/Controleur$1.class
#	target/classes/controleur/Controleur.class

+++ src/main/java/modele/Block.java
@@ -19,8 +19,8 @@ public class Block {
         this.x = x;
         this.y = y;
         this.tile = tile;
-        hitX = x + tile.getHitbox().getX();
-        hitY = y + tile.getHitbox().getY();
+        hitX = x + tile.getHitbox().getX().intValue();
+        hitY = y + tile.getHitbox().getY().intValue();
     }
 
     public int getX() {
